requirements.txt
Flask==2.1.0
flask-cors==3.0.10
newsapi-python==0.2.6
newspaper3k==0.2.8
nltk==3.6.5
requests==2.26.0